+++ dengr/CMakeLists.txt
@@ -2,8 +2,6 @@
 set(
     DENGR_SOURCES
     "src/ChannelFrame.cpp"
-    "src/eight_to_fourteen.cpp"
-    "src/scrambling.cpp"
 )
 # Header files
 set(
@@ -13,39 +11,26 @@ set(
     "include/ChannelByte.hpp"
     "include/ChannelFrame.hpp"
     "include/dengr.hpp"
-    "include/eight_to_fourteen.hpp"
+    "include/eight_to_fourteen/eight_to_fourteen.hpp"
     "include/F1Frame.hpp"
     "include/F2Frame.hpp"
     "include/F3Frame.hpp"
     "include/MergingBits.hpp"
     "include/Mode2Sector.hpp"
-    "include/physical_layer.hpp"
-    "include/physical_layer.inl"
+    "include/physical_layer/physical_layer.hpp"
+    "include/physical_layer/physical_layer.inl"
     "include/Pit.hpp"
     "include/ScrambledSector.hpp"
-    "include/scrambling.hpp"
+    "include/scrambling/scrambling.hpp"
     "include/Section.hpp"
 )
 
-# Generate rough (nearest major) version-dependent header installation folder
-set(
-    DENGR_ROUGH_HEADER_DESTINATION
-    "dengr-${PROJECT_VERSION_MAJOR}"
-)
-# Generate precise (major and minor) version-dependent header installation folder
-set(
-    DENGR_PRECISE_HEADER_DESTINATION
-    "dengr-${PROJECT_VERSION_MAJOR}.${PROJECT_VERSION_MINOR}"
-)
+add_subdirectory(src/eight_to_fourteen)
+add_subdirectory(src/scrambling)
 
 add_library(dengr)
 target_sources(dengr PRIVATE ${DENGR_SOURCES})
-target_include_directories(
-    dengr PUBLIC
-    $<BUILD_INTERFACE:${CMAKE_CURRENT_SOURCE_DIR}/include>
-    $<INSTALL_INTERFACE:include/${DENGR_PRECISE_HEADER_DESTINATION}>  # <prefix>/include/dengr-<version>
-)
-target_link_libraries(dengr PRIVATE dengr-compiler-options)
+target_link_libraries(dengr PRIVATE dengr-compiler-options dengr-efm dengr-scrambling)
 # set up version and soversion for the main library object
 set_target_properties(
     dengr PROPERTIES
