+++ dengr/include/scrambling/scrambling.hpp
