+++ dengr/src/eight_to_fourteen/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(dengr-efm eight_to_fourteen.cpp)
+target_link_libraries(dengr-efm PRIVATE dengr-compiler-options)
