+++ dengr/src/scrambling/CMakeLists.txt
@@ -1,0 +1,2 @@
+add_library(dengr-scrambling scrambling.cpp)
+target_link_libraries(dengr-scrambling PRIVATE dengr-compiler-options)
